requirements.txt
TgCrypto
pyrogram
python-dotenv